# docs/swagger.yaml
definitions:
...
  data.Function:
    properties:
      active:
        type: boolean
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      module:
        $ref: '#/definitions/data.Module'
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  data.Log:
    properties:
      createdAt:
        type: string
      detail:
        type: string
      id:
        type: string
      module:
        type: string
      params:
        type: string
      person_dni:
        type: string
      sensor:
        type: string
      user_nickname:
        type: string
    type: object
  data.Module:
    properties:
      functions:
        items:
          $ref: '#/definitions/data.Function'
        type: array
      name:
        type: string
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  routes.MessageResponse:
    properties:
      code:
        type: integer
      details:
        type: string
    type: object
...
  routes.logParams:
    properties:
      input:
        type: string
      output:
        type: string
      resource:
        type: string
    required:
    - input
    - output
    - resource
    type: object
...
  routes.paramsDevice:
    properties:
      datajson:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
  description: api autentia manager
  title: Autentia Manager Documentation
  version: "1"
paths:
...
  /device:
    get:
      consumes:
      - application/json
      description: get device
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/routes.MessageResponse'
      security:
      - BarerToken: []
      summary: get device
      tags:
      - device
    post:
      consumes:
      - application/json
      description: post device
      parameters:
      - description: device
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/routes.paramsDevice'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.MessageResponse'
      security:
      - BarerToken: []
      summary: post device
      tags:
      - device
...
  /logs/{date}:
    get:
      consumes:
      - application/json
      description: get logs
      parameters:
      - description: date
        in: path
        name: date
        required: true
        type: string
      - description: starthour
        in: query
        name: starthour
        required: true
        type: string
      - description: finishhour
        in: query
        name: finishhour
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/data.Log'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/routes.MessageResponse'
      security:
      - BarerToken: []
      summary: get Logs
      tags:
      - log
  /logs/trx:
    post:
      consumes:
      - application/json
      description: post logs
      parameters:
      - description: log
        in: body
        name: log
        required: true
        schema:
          $ref: '#/definitions/routes.logParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.MessageResponse'
      security:
      - BarerToken: []
      summary: post Logs
      tags:
      - log
  /modules:
    get:
      consumes:
      - application/json
      description: modules
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/data.Module'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/routes.MessageResponse'
      security:
      - BarerToken: []
      summary: modules
      tags:
      - module
  /modules/{name}:
    delete:
      consumes:
      - application/json
      description: delete modules
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.MessageResponse'
      security:
      - BarerToken: []
      summary: delete modules
      tags:
      - module
...
securityDefinitions:
  BarerToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"